Extract services fetcher as hook
